--- openstack/rds/v3/flavors/urls.go
@@ -1,9 +1,0 @@
-package flavors
-
-import "github.com/opentelekomcloud/gophertelekomcloud"
-
-const rootPath = "flavors"
-
-func listURL(client *golangsdk.ServiceClient, dbName string) string {
-	return client.ServiceURL(rootPath, dbName)
-}
